backend/requirements.txt
asgiref==3.7.2
boto3==1.28.78
celery==5.3.4
Django==4.2.10
django-cors-headers==4.2.0
django-filter==23.5
django-storages[boto3]==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
psycopg2-binary==2.9.9
pytest==7.4.4
pytest-cov==4.1.0
pytest-django==4.7.0
redis==5.0.1
requests==2.31.0
whitenoise==6.5.0
